airflow_crawling_test/requirements.txt
requests
beautifulsoup4
pandas
pytz
load_dotenv